created function for booking price calculation

+++ apps/hotel/models.py
@@ -24,7 +24,7 @@ class Booking(models.Model):
     roomtier = models.ForeignKey(RoomTier, on_delete=models.CASCADE)
     check_in = models.DateField()
     check_out = models.DateField()
-    total_price = models.IntegerField(max_digits=10, decimal_places=2, verbose_name=_('Total Price'))
+    total_price = models.DecimalField(max_digits=10, decimal_places=2, verbose_name=_('Total Price'))
 
     def __str__(self):
         return f"Booking of {self.roomtier.title} by {self.firstname} {self.lastname} from {self.check_in} to {self.check_out}"
